用户信息究极无敌大修改。文章右侧目录栏增加语音阅读

+++ blog/mojian-admin/src/main/java/com/mojian/service/SysUserService.java
@@ -32,7 +32,7 @@ public interface SysUserService extends IService<Users> {
     /**
      * 删除用户
      */
-    void delete(List<Integer> ids);
+    void delete(List<Long> ids);
 
 
     /**
